requirements_high_freq.txt
numpy>=1.21.0
pyserial>=3.5
scipy>=1.7.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0 